backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
et-xmlfile==1.1.0
idna==3.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mysqlclient==2.2.4
openpyxl==3.1.5
pip-chill==1.0.3
PyJWT==2.8.0
PyMySQL==1.1.1
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.31.0
rpds-py==0.20.0
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.2
